Still Nothing interesting :))))))

+++ srcs/maths/points_distance.c
@@ -6,5 +6,5 @@
 
 float	point_distance(t_coordinates p1, t_coordinates p2)
 {
-	return (sqrt(pow(p2.x - p1.x, 2) + pow(p2.y - p1.y, 2) + pow(p2.z - p1.z, 2)));
+	return (sqrtf(powf(p2.x - p1.x, 2) + powf(p2.y - p1.y, 2) + powf(p2.z - p1.z, 2)));
 }
